refactor: перенести картинки в ItGame
