+++ src/games/ItGame/components/Answer/Answer.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './Answer.css';
-import rightImage from '../../../../images/it-game_right.svg';
-import failImage from '../../../../images/it-game_fail.svg';
+import rightImage from '../../images/it-game_right.svg';
+import failImage from '../../images/it-game_fail.svg';
 
 const Answer = ({ answerOpened, isCorrectAnser }) => {
   return (
