+++ src/games/ItGame/pages/FinalPage/FinalPage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './FinalPage.css';
 import Header from '../../components/Header/Header';
-import finalEmoji from '../../../../images/it-game_final.svg';
+import finalEmoji from '../../images/it-game_final.svg';
 
 const FinalPage = () => {
   return (
